requirements.txt
fastapi
uvicorn
python-dotenv
appwrite